introduce battery page
